+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### [Unreleased]
+
+#### Fixed
+
+
 ### [0.0.0] - 2017-05-08
 
 #### Added