+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-### [Unreleased]
+### [0.0.1] - 2017-05-10
 
 #### Fixed
 
+* Allow bubbling through the DOM on click event by returning `true` instead of `false`.
+* Workaround for radio buttons: Remove attribute `"aria-pressed"` and class `active` manually to deactivate a radio button.
 
 ### [0.0.0] - 2017-05-08